requirements.txt
ultralytics>=8.0.20
opencv-python
pyttsx3
paho-mqtt
